Updated: Web Menu is now Routeable.

+++ app/Plugin/Dreamcms/Controller/WebMenusController.php
@@ -15,6 +15,13 @@ class WebMenusController extends DreamcmsAppController {
  */
 	public $components = array('Paginator');
 
+	public function beforeFilter()
+	{
+		parent::beforeFilter();
+
+		$this->Routeable->setParentModel($this->WebMenu);
+	}
+
 /**
  * index method
  *
@@ -30,6 +37,9 @@ class WebMenusController extends DreamcmsAppController {
 
 		$this->WebMenu->setLanguage(Configure::read('Config.language'));
 		$this->WebMenu->recursive = 0;
+		$paginate = $this->paginate;
+		$paginate['conditions'] = Set::merge(array('WebMenu.deleted' => 0), $this->Routeable->getFindConditions());
+		$this->paginate = $paginate;
 		$this->set('webMenus', $this->paginate());
 	}
 
@@ -43,11 +53,16 @@ class WebMenusController extends DreamcmsAppController {
 	public function view($id = null) {
 		$this->DreamcmsAcl->authorize();
 		if (!$this->WebMenu->exists($id)) {
-			throw new NotFoundException(__('Invalid web menu'));
+			throw new NotFoundException(__('Invalid ' . strtolower($this->Routeable->singularize)));
 		}
 		$this->WebMenu->setLanguage(Configure::read('Config.language'));
-		$options = array('conditions' => array('WebMenu.' . $this->WebMenu->primaryKey => $id));
-		$this->set('webMenu', $this->WebMenu->find('first', $options));
+		$conditions = Set::merge(array('WebMenu.' . $this->WebMenu->primaryKey => $id, 'WebMenu.deleted' => 0), $this->Routeable->getFindConditions());
+		$options = array('conditions' => $conditions);
+		$webMenu = $this->WebMenu->find('first', $options);
+		if (!$webMenu) {
+			throw new NotFoundException(__('Invalid ' . strtolower($this->Routeable->singularize)));
+		}
+		$this->set('webMenu', $webMenu);
 	}
 
 /**
@@ -62,11 +77,12 @@ class WebMenusController extends DreamcmsAppController {
 			$this->Translator->init($this->WebMenu, $this->request->data);
 			if ($this->Translator->validate()) {
 				$this->Translator->save();
-				$this->Session->setFlash(__('The web menu has been saved'), 'flash/success');
-				$this->redirect(array('action' => 'index'));
+				$this->Session->setFlash(__('The '. strtolower($this->Routeable->singularize) .' has been saved'), 'flash/success');
+				$this->redirect(array('controller' => $this->Routeable->currentController, 'action' => 'index'));
 			}
 		}
-		$this->set('parentWebMenus', $this->WebMenu->generateTreeList());
+		$parentWebMenus = $this->WebMenu->generateTreeList(Set::merge(array('WebMenu.deleted' => 0), $this->Routeable->getTreeListConditions()));
+		$this->set('parentWebMenus', $parentWebMenus);
 	}
 
 /**
@@ -80,22 +96,24 @@ class WebMenusController extends DreamcmsAppController {
 		$this->DreamcmsAcl->authorize();
         $this->WebMenu->id = $id;
 		if (!$this->WebMenu->exists($id)) {
-			throw new NotFoundException(__('Invalid web menu'));
+			throw new NotFoundException(__('Invalid ' . strtolower($this->Routeable->singularize)));
 		}
 		$this->WebMenu->setLanguage(Configure::read('Config.language'));
 		if ($this->request->is('post') || $this->request->is('put')) {
 			$this->Translator->init($this->WebMenu, $this->request->data);
 			if ($this->Translator->validate()) {
 				$this->Translator->save();
-				$this->Session->setFlash(__('The web menu has been saved'), 'flash/success');
-				$this->redirect(array('action' => 'index'));
+				$this->Session->setFlash(__('The '. strtolower($this->Routeable->singularize) .' has been saved'), 'flash/success');
+				$this->redirect(array('controller' => $this->Routeable->currentController, 'action' => 'index'));
 			}
 		} else {
-			$options = array('conditions' => array('WebMenu.' . $this->WebMenu->primaryKey => $id));
-			//$this->request->data = $this->WebMenu->find('first', $options);
+			$options = array('conditions' => Set::merge(array('WebMenu.' . $this->WebMenu->primaryKey => $id, 'WebMenu.deleted' => 0), $this->Routeable->getFindConditions()));
 			$this->request->data = $this->Translator->findFirst($this->WebMenu, $options);
+			if (!$this->request->data)
+				throw new NotFoundException(__('Invalid ' . strtolower($this->Routeable->singularize)));
 		}
-		$this->set('parentWebMenus', $this->WebMenu->generateTreeList());
+		$parentWebMenus = $this->WebMenu->generateTreeList(Set::merge(array('WebMenu.deleted' => 0), $this->Routeable->getTreeListConditions()));
+		$this->set('parentWebMenus', $parentWebMenus);
 	}
 
 /**
@@ -113,12 +131,17 @@ class WebMenusController extends DreamcmsAppController {
 		}
 		$this->WebMenu->id = $id;
 		if (!$this->WebMenu->exists()) {
-			throw new NotFoundException(__('Invalid web menu'));
+			throw new NotFoundException(__('Invalid ' . strtolower($this->Routeable->singularize)));
 		}
+		$conditions = Set::merge(array('WebMenu.deleted' => 0), $this->Routeable->getFindConditions());
+		$webMenu = $this->Translator->findFirst($this->WebMenu, array('conditions' => $conditions));
+		if (!$webMenu)
+			throw new NotFoundException(__('Invalid ' . strtolower($this->Routeable->singularize)));
+
 		if ($this->WebMenu->delete()) {
-			$this->Session->setFlash(__('Web menu deleted'), 'flash/success');
-			$this->redirect(array('action' => 'index'));
+			$this->Session->setFlash(__($this->Routeable->singularize . ' deleted'), 'flash/success');
+			$this->redirect(array('controller' => $this->Routeable->currentController, 'action' => 'index'));
 		}
-		$this->Session->setFlash(__('Web menu was not deleted'), 'flash/error');
-		$this->redirect(array('action' => 'index'));
+		$this->Session->setFlash(__($this->Routeable->singularize . ' was not deleted'), 'flash/error');
+		$this->redirect(array('controller' => $this->Routeable->currentController, 'action' => 'index'));
 	}}
